backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
SQLAlchemy>=2.0.32
greenlet>=3.0.0
aiosqlite>=0.20.0
PyGithub>=2.4.0
google-generativeai>=0.8.3
openai>=1.44.0
httpx>=0.27.0
qdrant-client>=1.11.0
python-dotenv>=1.0.1
